# openapi.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - App
  /users:
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Users
  /users/{id}:
    get:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
      tags:
        - Users
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: ''
      tags:
        - Users
    delete:
      operationId: UsersController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags:
        - Users
info:
  title: Tasks
  description: The users API description
  version: '1.0'
  contact: {}
tags:
  - name: tasks
    description: ''
servers: []
components:
  schemas:
    User:
      type: object
      properties: {}
    UserDto:
      type: object
      properties:
        name:
          type: string
          description: User name
          minLength: 3
          maxLength: 30
          default: Default name
        email:
          type: string
          description: email
          minLength: 11
          maxLength: 44
          default: Default email
      required:
        - name
        - email